+++ RELEASES.md
@@ -563,3 +563,13 @@ Based on:
 - [typescript v5.2.0-alpha.16] .
 ### Releases
 - [NPM v5.2.0-alpha.16] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.16 - .
+
+## 2025-03-18 15:54:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.519.0 (2.552.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.17] .
+### Releases
+- [NPM v5.2.0-alpha.17] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.17 - .